Этап 3

+++ analyzer/src/main/java/ru/practicum/analyzer/grpc/RecommendationsControllerGrpcImpl.java
@@ -75,7 +75,7 @@ public class RecommendationsControllerGrpcImpl
 		for (RecommendedEvent e : events) {
 			responseObserver.onNext(RecommendedEventProto.newBuilder()
 					.setEventId(e.eventId())
-					.setScore((float) e.score())
+					.setScore(e.score())
 					.build());
 		}
 		responseObserver.onCompleted();
